requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
chromadb>=1.0.0
flask>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
Pillow>=10.0.0
